% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mine}
\alias{mine}
\title{Mine}
\usage{
mine(x, y, alpha = 0.6, c = 15)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{another numeric vector}

\item{alpha}{exponent on (0,1)}

\item{c}{number of clumps}
}
\description{
Calculate the MIC.
}
\examples{
data("iris")
mine(iris$Sepal.Length, iris$Sepal.Width)

}

